# trading-service/docs/swagger.yaml
basePath: /
definitions:
...
  types.CreateOrderRequest:
    properties:
      account_id:
        type: integer
      aon:
        type: boolean
      contract_size:
        type: integer
      direction:
        type: string
      limit_price_per_unit:
        type: number
      margin:
        type: boolean
      quantity:
        type: integer
      security_id:
        type: integer
      stop_price_per_unit:
        type: number
      user_id:
        type: integer
    required:
    - account_id
    - aon
    - contract_size
    - direction
    - margin
    - quantity
    - security_id
    - user_id
    type: object
...
  types.OrderResponse:
    properties:
      account_id:
        type: integer
      after_hours:
        type: boolean
      aon:
        type: boolean
      approved_by:
        description: Supervizor koji je odobrio order
        type: integer
      contract_size:
        type: integer
      direction:
        type: string
      id:
        type: integer
      is_done:
        type: boolean
      last_modified:
        type: integer
      limit_price_per_unit:
        type: number
      margin:
        type: boolean
      quantity:
        type: integer
      remaining_parts:
        type: integer
      security_id:
        type: integer
      status:
        type: string
      stop_price_per_unit:
        type: number
      user_id:
        type: integer
    type: object
  types.Response:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: Trading Service API
  title: Trading Service
  version: "1.0"
paths:
...
  /orders:
    get:
      description: Vraća listu naloga, opciono filtriranu po statusu.
      parameters:
      - default: all
        description: Status naloga za filtriranje. Podrazumevano 'all' za sve statuse.
        example: pending
        in: query
        name: filter_status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Uspešno preuzeta lista naloga
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/types.OrderResponse'
                  type: array
              type: object
        "500":
          description: Greška pri preuzimanju naloga iz baze
          schema:
            $ref: '#/definitions/types.Response'
      summary: Preuzimanje liste naloga
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Kreira novi nalog za hartije od vrednosti.
      parameters:
      - description: Podaci neophodni za kreiranje naloga
        in: body
        name: orderRequest
        required: true
        schema:
          $ref: '#/definitions/types.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Uspešno kreiran nalog, vraća ID novog naloga
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: Neispravan format, neuspela validacija ili greška pri upisu
            u bazu
          schema:
            $ref: '#/definitions/types.Response'
        "403":
          description: Nije dozvoljeno kreirati nalog za drugog korisnika
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Kreiranje novog naloga
      tags:
      - Orders
  /orders/{id}:
    get:
      description: Vraća detalje specifičnog naloga na osnovu njegovog jedinstvenog
        identifikatora (ID).
      parameters:
      - description: ID naloga koji se preuzima
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Uspešno preuzet nalog
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/types.OrderResponse'
              type: object
        "400":
          description: Nevalidan ID naloga
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Nalog sa datim ID-jem ne postoji
          schema:
            $ref: '#/definitions/types.Response'
      summary: Preuzimanje naloga po ID-u
      tags:
      - Orders
  /orders/{id}/approve:
    post:
      description: Menja status naloga u 'approved'.
      parameters:
      - description: ID naloga koji se odobrava
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Nalog uspešno odobren
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: Nevalidan ID ili nalog nije u 'pending' statusu
          schema:
            $ref: '#/definitions/types.Response'
        "403":
          description: Nedovoljne privilegije
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Nalog sa datim ID-jem ne postoji
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Interna Greška Servera
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Odobravanje naloga
      tags:
      - Orders
  /orders/{id}/decline:
    post:
      description: Menja status naloga u 'declined'.
      parameters:
      - description: ID naloga koji se odbija
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Nalog uspešno odbijen, vraća ID naloga
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: Nevalidan ID ili nalog nije u 'pending' statusu
          schema:
            $ref: '#/definitions/types.Response'
        "403":
          description: Nedovoljne privilegije
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Nalog sa datim ID-jem ne postoji
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Interna Greška Servera
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Odbijanje naloga
      tags:
      - Orders
...
securityDefinitions:
  BearerAuth:
    description: 'Unesite token. Primer: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"